Rename data stream transport actions (#127222)
The new action names are more consistent with the rest of the codebase.

(cherry picked from commit 4207cee3eb10f6e250234acbc4bafcfca785b044)

# Conflicts:
#	modules/data-streams/src/main/java/org/elasticsearch/datastreams/DataStreamsPlugin.java
#	modules/data-streams/src/main/java/org/elasticsearch/datastreams/action/TransportMigrateToDataStreamAction.java
#	modules/data-streams/src/main/java/org/elasticsearch/datastreams/action/TransportModifyDataStreamsAction.java
#	modules/data-streams/src/main/java/org/elasticsearch/datastreams/action/TransportPromoteDataStreamAction.java
#	modules/data-streams/src/test/java/org/elasticsearch/datastreams/action/TransportDeleteDataStreamActionTests.java
