+++ modules/data-streams/src/main/java/org/elasticsearch/datastreams/DataStreamsPlugin.java
@@ -34,13 +34,13 @@ import org.elasticsearch.common.settings.Settings;
 import org.elasticsearch.common.settings.SettingsFilter;
 import org.elasticsearch.core.IOUtils;
 import org.elasticsearch.core.TimeValue;
-import org.elasticsearch.datastreams.action.CreateDataStreamTransportAction;
-import org.elasticsearch.datastreams.action.DataStreamsStatsTransportAction;
-import org.elasticsearch.datastreams.action.DeleteDataStreamTransportAction;
-import org.elasticsearch.datastreams.action.MigrateToDataStreamTransportAction;
-import org.elasticsearch.datastreams.action.ModifyDataStreamsTransportAction;
-import org.elasticsearch.datastreams.action.PromoteDataStreamTransportAction;
+import org.elasticsearch.datastreams.action.TransportCreateDataStreamAction;
+import org.elasticsearch.datastreams.action.TransportDataStreamsStatsAction;
+import org.elasticsearch.datastreams.action.TransportDeleteDataStreamAction;
 import org.elasticsearch.datastreams.action.TransportGetDataStreamsAction;
+import org.elasticsearch.datastreams.action.TransportMigrateToDataStreamAction;
+import org.elasticsearch.datastreams.action.TransportModifyDataStreamsAction;
+import org.elasticsearch.datastreams.action.TransportPromoteDataStreamAction;
 import org.elasticsearch.datastreams.lifecycle.DataStreamLifecycleErrorStore;
 import org.elasticsearch.datastreams.lifecycle.DataStreamLifecycleService;
 import org.elasticsearch.datastreams.lifecycle.action.DeleteDataStreamLifecycleAction;
@@ -227,13 +227,13 @@ public class DataStreamsPlugin extends Plugin implements ActionPlugin, HealthPlu
     @Override
     public List<ActionHandler<? extends ActionRequest, ? extends ActionResponse>> getActions() {
         List<ActionHandler<? extends ActionRequest, ? extends ActionResponse>> actions = new ArrayList<>();
-        actions.add(new ActionHandler<>(CreateDataStreamAction.INSTANCE, CreateDataStreamTransportAction.class));
-        actions.add(new ActionHandler<>(DeleteDataStreamAction.INSTANCE, DeleteDataStreamTransportAction.class));
+        actions.add(new ActionHandler<>(CreateDataStreamAction.INSTANCE, TransportCreateDataStreamAction.class));
+        actions.add(new ActionHandler<>(DeleteDataStreamAction.INSTANCE, TransportDeleteDataStreamAction.class));
         actions.add(new ActionHandler<>(GetDataStreamAction.INSTANCE, TransportGetDataStreamsAction.class));
-        actions.add(new ActionHandler<>(DataStreamsStatsAction.INSTANCE, DataStreamsStatsTransportAction.class));
-        actions.add(new ActionHandler<>(MigrateToDataStreamAction.INSTANCE, MigrateToDataStreamTransportAction.class));
-        actions.add(new ActionHandler<>(PromoteDataStreamAction.INSTANCE, PromoteDataStreamTransportAction.class));
-        actions.add(new ActionHandler<>(ModifyDataStreamsAction.INSTANCE, ModifyDataStreamsTransportAction.class));
+        actions.add(new ActionHandler<>(DataStreamsStatsAction.INSTANCE, TransportDataStreamsStatsAction.class));
+        actions.add(new ActionHandler<>(MigrateToDataStreamAction.INSTANCE, TransportMigrateToDataStreamAction.class));
+        actions.add(new ActionHandler<>(PromoteDataStreamAction.INSTANCE, TransportPromoteDataStreamAction.class));
+        actions.add(new ActionHandler<>(ModifyDataStreamsAction.INSTANCE, TransportModifyDataStreamsAction.class));
         actions.add(new ActionHandler<>(PutDataStreamLifecycleAction.INSTANCE, TransportPutDataStreamLifecycleAction.class));
         actions.add(new ActionHandler<>(GetDataStreamLifecycleAction.INSTANCE, TransportGetDataStreamLifecycleAction.class));
         actions.add(new ActionHandler<>(DeleteDataStreamLifecycleAction.INSTANCE, TransportDeleteDataStreamLifecycleAction.class));
