+++ modules/data-streams/src/main/java/org/elasticsearch/datastreams/action/TransportDataStreamsStatsAction.java
@@ -48,7 +48,7 @@ import java.util.Map;
 import java.util.Set;
 import java.util.SortedMap;
 
-public class DataStreamsStatsTransportAction extends TransportBroadcastByNodeAction<
+public class TransportDataStreamsStatsAction extends TransportBroadcastByNodeAction<
     DataStreamsStatsAction.Request,
     DataStreamsStatsAction.Response,
     DataStreamsStatsAction.DataStreamShardStats> {
@@ -56,7 +56,7 @@ public class DataStreamsStatsTransportAction extends TransportBroadcastByNodeAct
     private final IndicesService indicesService;
 
     @Inject
-    public DataStreamsStatsTransportAction(
+    public TransportDataStreamsStatsAction(
         ClusterService clusterService,
         TransportService transportService,
         IndicesService indicesService,
